add conditional logic for filtering rows
This allows one to require a certain column to be populated only when a
specified value is found in another column.  This logic can be extended
to the controlled, date, and number rules (it is not needed for the
required and allowed rules, which test only the existence of columns).